// apps/web-solid-next/src/__generated__/graphql-types.ts
export enum Rating {
  Dislike = 'DISLIKE',
  Like = 'LIKE'
}